change id in check_death to a long like the struct

+++ philo/main.c
@@ -6,7 +6,7 @@
 /*   By: croy <croy@student.42lyon.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 14:32:09 by croy              #+#    #+#             */
-/*   Updated: 2023/03/24 14:26:02 by croy             ###   ########lyon.fr   */
+/*   Updated: 2023/03/27 14:06:58 by croy             ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,7 +88,7 @@ void	*philo_routine(void *arg)
  */
 int	check_death(t_data *data, t_philo *philos)
 {
-	int		id;
+	long	id;
 	long	since_meal;
 
 	while (data->all_alive)
@@ -96,7 +96,6 @@ int	check_death(t_data *data, t_philo *philos)
 		id = 0;
 		while(id < data->nb_philo)
 		{
-			// if (philo->data->meal_limit > 0 && philo->eaten >= philo->data->meal_limit)
 			if (data->need_to_eat == 0)
 				return (0);
 			since_meal = get_time(philos[id].last_meal);
